Add Telegram Bot

+++ backend/app/models/cart.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, Float, ForeignKey
+from sqlalchemy import Integer, BigInteger, Float, ForeignKey
 from sqlalchemy.orm import relationship, Mapped, mapped_column
 
 from database import Base
@@ -8,7 +8,8 @@ class CartItem(Base):
     __tablename__ = "cart_items"
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True)
-    chat_id: Mapped[int] = mapped_column(Integer, ForeignKey("users.chat_id"))
+    chat_id: Mapped[int] = mapped_column(BigInteger,
+                                         ForeignKey("users.chat_id"))
     product_id: Mapped[int] = mapped_column(Integer, ForeignKey("products.id"))
     quantity: Mapped[int] = mapped_column(Integer, default=1)
     total_price: Mapped[float] = mapped_column(Float, nullable=True)
